+++ packages/ledger-amino/src/launchpadledger.ts
@@ -1,9 +1,10 @@
-import { makeCosmoshubPath } from "@cosmjs/amino";
+import { encodeSecp256k1Pubkey, makeCosmoshubPath, pubkeyToAddress, Secp256k1Pubkey } from "@cosmjs/amino";
 import { HdPath, Secp256k1Signature } from "@cosmjs/crypto";
 import { fromUtf8 } from "@cosmjs/encoding";
 import { assert } from "@cosmjs/utils";
 import Transport from "@ledgerhq/hw-transport";
 import CosmosApp, {
+  AddressAndPublicKeyResponse,
   AppInfoResponse,
   PublicKeyResponse,
   SignResponse,
@@ -49,6 +50,11 @@ export interface LaunchpadLedgerOptions {
    * Defaults to "1.5.3".
    */
   readonly minLedgerAppVersion?: string;
+}
+
+export interface AddressAndPubkey {
+  readonly address: string;
+  readonly pubkey: Secp256k1Pubkey;
 }
 
 export class LaunchpadLedger {
@@ -109,7 +115,7 @@ export class LaunchpadLedger {
 
   public async getCosmosAddress(pubkey?: Uint8Array): Promise<string> {
     const pubkeyToUse = pubkey || (await this.getPubkey());
-    return CosmosApp.getBech32FromPK(this.prefix, Buffer.from(pubkeyToUse));
+    return pubkeyToAddress(encodeSecp256k1Pubkey(pubkeyToUse), this.prefix);
   }
 
   public async sign(message: Uint8Array, hdPath?: HdPath): Promise<Uint8Array> {
@@ -164,6 +170,29 @@ export class LaunchpadLedger {
   private async verifyDeviceIsReady(): Promise<void> {
     await this.verifyAppVersion();
     await this.verifyCosmosAppIsOpen();
+  }
+
+  /**
+   * Shows the user's address in the device and returns an address/pubkey pair.
+   *
+   * The address will be shown with the native prefix of the app (e.g. cosmos, persistence, desmos)
+   * and does not support the usage of other address prefixes.
+   *
+   * @param path The HD path to show the address for. If unset, this is the first account.
+   */
+  public async showAddress(hdPath?: HdPath): Promise<AddressAndPubkey> {
+    await this.verifyDeviceIsReady();
+
+    const hdPathToUse = hdPath || this.hdPaths[0];
+    // ledger-cosmos-js hardens the first three indices
+    const response = await this.app.showAddressAndPubKey(unharden(hdPathToUse), this.prefix);
+    this.handleLedgerErrors(response);
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    const { address, compressed_pk } = response as AddressAndPublicKeyResponse;
+    return {
+      address: address,
+      pubkey: encodeSecp256k1Pubkey(compressed_pk),
+    };
   }
 
   private handleLedgerErrors(