+++ packages/ledger-amino/src/ledgersigner.ts
@@ -10,7 +10,7 @@ import {
 import { HdPath } from "@cosmjs/crypto";
 import Transport from "@ledgerhq/hw-transport";
 
-import { LaunchpadLedger, LaunchpadLedgerOptions } from "./launchpadledger";
+import { AddressAndPubkey, LaunchpadLedger, LaunchpadLedgerOptions } from "./launchpadledger";
 
 export class LedgerSigner implements OfflineAminoSigner {
   private readonly ledger: LaunchpadLedger;
@@ -35,6 +35,18 @@ export class LedgerSigner implements OfflineAminoSigner {
     }
 
     return this.accounts;
+  }
+
+  /**
+   * Shows the user's address in the device and returns an address/pubkey pair.
+   *
+   * The address will be shown with the native prefix of the app (e.g. cosmos, persistence, desmos)
+   * and does not support the usage of other address prefixes.
+   *
+   * @param path The HD path to show the address for. If unset, this is the first account.
+   */
+  public async showAddress(path?: HdPath): Promise<AddressAndPubkey> {
+    return this.ledger.showAddress(path);
   }
 
   public async signAmino(signerAddress: string, signDoc: StdSignDoc): Promise<AminoSignResponse> {