+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
   JavaScript objects and the JSON representation of `cosmwasm_std::Binary`
   (base64).
 - @cosmjs/cosmwasm-stargate: Export `WasmExtension` and `setupWasmExtension`.
+- @cosmjs/ledger-amino: Added `LedgerSigner.showAddress` and
+  `LaunchpadLedger.showAddress` to show the user's address in the Ledger screen.
 
 ### Changed